[ Modify ] Separate Components Folder
