+++ front/components/dashboard/DailyActivity.js
